+++ libfreeipmi/include/freeipmi/interface/ipmi-rmcpplus-interface.h
@@ -254,22 +254,22 @@ int32_t assemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
                                     uint8_t *pkt,
                                     uint32_t pkt_len);
 
-int32_t unassemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
-                                      uint8_t integrity_algorithm,
-                                      uint8_t confidentiality_algorithm,
-                                      uint8_t *integrity_key,
-                                      uint32_t integrity_key_len,
-                                      uint8_t *confidentiality_key,
-                                      uint32_t confidentiality_key_len,
-                                      uint8_t *pkt,
-                                      uint32_t pkt_len,
-                                      fiid_obj_t obj_rmcp_hdr,
-                                      fiid_obj_t obj_rmcpplus_session_hdr,
-                                      fiid_obj_t obj_rmcpplus_payload,
-                                      fiid_obj_t obj_lan_msg_hdr,
-                                      fiid_obj_t obj_cmd,
-                                      fiid_obj_t obj_lan_msg_trlr,
-                                      fiid_obj_t obj_rmcpplus_session_trlr);
+int8_t unassemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
+                                     uint8_t integrity_algorithm,
+                                     uint8_t confidentiality_algorithm,
+                                     uint8_t *integrity_key,
+                                     uint32_t integrity_key_len,
+                                     uint8_t *confidentiality_key,
+                                     uint32_t confidentiality_key_len,
+                                     uint8_t *pkt,
+                                     uint32_t pkt_len,
+                                     fiid_obj_t obj_rmcp_hdr,
+                                     fiid_obj_t obj_rmcpplus_session_hdr,
+                                     fiid_obj_t obj_rmcpplus_payload,
+                                     fiid_obj_t obj_lan_msg_hdr,
+                                     fiid_obj_t obj_cmd,
+                                     fiid_obj_t obj_lan_msg_trlr,
+                                     fiid_obj_t obj_rmcpplus_session_trlr);
 
 #ifdef __cplusplus
 }
